Refactor functionality into several core submodules
